requirements.txt
google-genai>=0.2.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydub>=0.25.1
numpy>=1.24.0
pydantic>=2.0.0
rank-bm25>=0.2.2
pypdf2>=3.0.0
schedule>=1.2.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
